pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chucks"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
# license = None
keywords = []
authors = [
  { name = "John Fiocca", email = "fioccajohn@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
	"schwab-py"
]

[project.urls]
Documentation = "https://github.com/fioccajohn/chucks#readme"
Issues = "https://github.com/fioccajohn/chucks/issues"
Source = "https://github.com/fioccajohn/chucks"

[tool.hatch.version]
path = "src/chucks/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/chucks tests}"

[tool.coverage.run]
source_pkgs = ["chucks", "tests"]
branch = true
parallel = true
omit = [
  "src/chucks/__about__.py",
]

[tool.coverage.paths]
chucks = ["src/chucks", "*/chucks/src/chucks"]
tests = ["tests", "*/chucks/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
